Do not attempt to cleanup SelfSubjectAccessReview

+++ tests/framework/clients/dynamic/dynamic.go
@@ -72,6 +72,25 @@ type ResourceClient struct {
 	ts *session.Session
 }
 
+var (
+	noCleanupGVKs = []schema.GroupVersionKind{
+		{
+			Group:   "authorization.k8s.io",
+			Version: "v1",
+			Kind:    "SelfSubjectAccessReview",
+		},
+	}
+)
+
+func needsCleanup(obj *unstructured.Unstructured) bool {
+	for _, gvk := range noCleanupGVKs {
+		if obj.GroupVersionKind() == gvk {
+			return false
+		}
+	}
+	return true
+}
+
 // Create is dynamic.ResourceInterface's Create function, that is being overwritten to register its delete function to the session.Session
 // that is being reference.
 func (c *ResourceClient) Create(ctx context.Context, obj *unstructured.Unstructured, opts metav1.CreateOptions, subresources ...string) (*unstructured.Unstructured, error) {
@@ -80,20 +99,22 @@ func (c *ResourceClient) Create(ctx context.Context, obj *unstructured.Unstructu
 		return nil, err
 	}
 
-	c.ts.RegisterCleanupFunc(func() error {
-		err := c.Delete(context.TODO(), unstructuredObj.GetName(), metav1.DeleteOptions{}, subresources...)
-		if errors.IsNotFound(err) {
-			return nil
-		}
+	if needsCleanup(obj) {
+		c.ts.RegisterCleanupFunc(func() error {
+			err := c.Delete(context.TODO(), unstructuredObj.GetName(), metav1.DeleteOptions{}, subresources...)
+			if errors.IsNotFound(err) {
+				return nil
+			}
 
-		name := unstructuredObj.GetName()
-		if unstructuredObj.GetNamespace() != "" {
-			name = unstructuredObj.GetNamespace() + "/" + name
-		}
-		gvk := unstructuredObj.GetObjectKind().GroupVersionKind()
+			name := unstructuredObj.GetName()
+			if unstructuredObj.GetNamespace() != "" {
+				name = unstructuredObj.GetNamespace() + "/" + name
+			}
+			gvk := unstructuredObj.GetObjectKind().GroupVersionKind()
 
-		return fmt.Errorf("%v (%v): %w", name, gvk, err)
-	})
+			return fmt.Errorf("%v (%v): %w", name, gvk, err)
+		})
+	}
 
 	return unstructuredObj, err
 }
